py: document how to write tests with a shared pipeline
