+++ python/tests/__init__.py
@@ -1,5 +1,6 @@
 from feldera.rest import FelderaClient
 import os
+import unittest
 
 BASE_URL = os.environ.get("FELDERA_BASE_URL", "http://localhost:8080")
 KAFKA_SERVER = os.environ.get("FELDERA_KAFKA_SERVER", "localhost:19092")
@@ -8,3 +9,11 @@ PIPELINE_TO_KAFKA_SERVER = os.environ.get(
 )
 
 TEST_CLIENT = FelderaClient(BASE_URL)
+
+
+def enterprise_only(fn):
+    fn._enterprise_only = True
+    return unittest.skipUnless(
+        TEST_CLIENT.get_config().edition.is_enterprise(),
+        f"{fn.__name__} is enterprise only, skipping",
+    )(fn)
